requirements.txt
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai
networkx
pyvis
aiohttp
pyyaml
websockets
httpx
aiohttp
weasyprint
markdown